# internal/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_kimnt93_gorouter_internal_api.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorDetail'
    type: object
...
  internal_api_handlers.APIKeyCreateRequest:
    properties:
      context_organization_id:
        type: string
      models:
        items:
          type: string
        type: array
      name:
        type: string
      owner_organization_id:
        type: string
      owner_type:
        type: string
      owner_user_id:
        type: string
      quota_period:
        type: string
      quota_usd:
        type: number
      scopes:
        items:
          type: string
        type: array
      tenant_id:
        type: string
      workload:
        $ref: '#/definitions/github_com_kimnt93_gorouter_pkg_entities.WorkloadBinding'
    type: object
  internal_api_handlers.APIKeyListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_kimnt93_gorouter_pkg_entities.ApiKey'
        type: array
      next_cursor:
        type: string
      object:
        type: string
    type: object
...
  internal_api_handlers.APIKeyRevealResponse:
    properties:
      plaintext:
        type: string
    type: object
...
  internal_api_handlers.OKResponse:
    properties:
      ok:
        type: boolean
    type: object
...
info:
  contact: {}
  description: Principal-aware model gateway and administration API.
  title: gorouter API
  version: "1.0"
paths:
  /admin/api-keys:
    get:
      description: Lists API-key metadata after applying principal ownership, organization
        context, status, and pagination filters.
      parameters:
      - description: user or organization
        in: query
        name: owner_type
        type: string
      - description: Owner user or organization ID
        in: query
        name: owner_id
        type: string
      - description: Context organization ID
        in: query
        name: organization_id
        type: string
      - description: Master-only user View As filter
        in: query
        name: view_user_id
        type: string
      - description: enabled or disabled
        in: query
        name: status
        type: string
      - description: Opaque cursor
        in: query
        name: cursor
        type: string
      - default: 100
        description: Page size
        in: query
        maximum: 500
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.APIKeyListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List API keys
      tags:
      - api-keys
    post:
      description: Creates an owned API key and returns its plaintext secret exactly
        once in the response.
      parameters:
      - description: Key configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.APIKeyCreateRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api_handlers.CreatedAPIKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create an API key
      tags:
      - api-keys
  /admin/api-keys/{id}:
    delete:
      description: Permanently deletes an API key after checking ownership and management
        permission.
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.OKResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an API key
      tags:
      - api-keys
    patch:
      description: Changes mutable API-key policy fields such as enabled state, model
        allowlist, scopes, quota, and RPM.
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      - description: Mutable key fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.APIKeyPatchRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.OKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an API key
      tags:
      - api-keys
  /admin/api-keys/{id}/reveal:
    get:
      description: Reveals an API key to master, its owner, or an administrator of
        its context organization.
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.APIKeyRevealResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reveal an API key
      tags:
      - api-keys
  /admin/api-keys/{id}/rotate:
    post:
      description: Replaces an API key secret and returns the new plaintext secret
        exactly once.
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.CreatedAPIKeyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Rotate an API key
      tags:
      - api-keys
  /admin/api-keys/models:
    get:
      description: Lists models that the current principal may assign to an API key,
        including effective pricing.
      parameters:
      - description: Target organization
        in: query
        name: organization_id
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.APIKeyModelOptionsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List grantable API-key models
      tags:
      - api-keys
  /admin/audit/events:
    get:
      description: Returns policy-constrained, cursor-paginated administrative audit
        events.
      parameters:
      - description: Opaque cursor
        in: query
        name: cursor
        type: string
      - default: 100
        description: Page size
        in: query
        maximum: 500
        name: limit
        type: integer
      - description: RFC3339 lower time bound
        in: query
        name: since
        type: string
      - description: RFC3339 upper time bound
        in: query
        name: until
        type: string
      - description: Organization filter
        in: query
        name: organization_id
        type: string
      - description: Master-only user View As filter
        in: query
        name: view_user_id
        type: string
      - description: Actor ID
        in: query
        name: actor_id
        type: string
      - description: Action
        in: query
        name: action
        type: string
      - description: Target type
        in: query
        name: target_type
        type: string
      - description: Target ID
        in: query
        name: target_id
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.AuditListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List audit events
      tags:
      - audit
  /admin/cache/flush:
    post:
      description: Purges deterministic router response-cache entries.
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.OKResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Flush prompt cache
      tags:
      - cache
  /admin/cache/stats:
    get:
      description: Returns deterministic router response-cache counters without exposing
        cache contents.
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_pkg_chat.CacheStats'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get cache statistics
      tags:
      - cache
  /admin/capabilities:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.CapabilitiesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get dependency capabilities
      tags:
      - usage
  /admin/credentials:
    get:
      description: Lists safe credential metadata or creates an encrypted provider
        credential for the authorized owner or organization.
      parameters:
      - description: Organization context
        in: query
        name: organization_id
        type: string
      - description: Master-only user View As filter
        in: query
        name: view_user_id
        type: string
      - description: Required for POST
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_api_handlers.CredentialCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_api_handlers.CredentialResponse'
            type: array
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api_handlers.CredentialResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List or create credentials
      tags:
      - credentials
    post:
      description: Lists safe credential metadata or creates an encrypted provider
        credential for the authorized owner or organization.
      parameters:
      - description: Organization context
        in: query
        name: organization_id
        type: string
      - description: Master-only user View As filter
        in: query
        name: view_user_id
        type: string
      - description: Required for POST
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_api_handlers.CredentialCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_api_handlers.CredentialResponse'
            type: array
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_api_handlers.CredentialResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List or create credentials
      tags:
      - credentials
...
  /admin/credentials/{id}/models:
    get:
      description: Discovers safe model metadata through a provider credential.
      parameters:
      - description: Credential ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.ProviderModelsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Discover provider models
      tags:
      - credentials
...
  /admin/session:
    get:
      description: Returns safe metadata for the current authenticated browser or
        bearer session.
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kimnt93_gorouter_internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current session
      tags:
      - authentication
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Use "Bearer {master-or-api-key}".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"